Extract LMS color space into sub-module
